requirements.txt
# Core RAG Framework
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.13
langchain-chroma==0.1.4
langchain-text-splitters==0.3.4
langchain-groq

# Vector Database
chromadb==1.3.5

# Document Processing
pypdf==5.1.0

# FastAPI Server
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20

# Environment Variables
python-dotenv==1.0.1

# OpenAI
openai==1.58.1

# Additional Utilities
pydantic==2.10.4
pydantic-settings==2.7.0
